Refactor webpack config into a merged model
